Refactor HPO default embedding range
