+++ src/pykeen/models/unimodal/hole.py
@@ -8,7 +8,7 @@ import torch
 import torch.autograd
 
 from ..base import EntityRelationEmbeddingModel
-from ...hpo.search_spaces import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
+from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
 from ...losses import Loss
 from ...nn.init import xavier_uniform_
 from ...regularizers import Regularizer
