+++ src/pykeen/models/unimodal/kg2e.py
@@ -9,7 +9,7 @@ import torch
 import torch.autograd
 
 from ..base import EntityRelationEmbeddingModel
-from ...hpo.search_spaces import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
+from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
 from ...losses import Loss
 from ...nn import Embedding
 from ...regularizers import Regularizer
