+++ src/pykeen/models/unimodal/ntn.py
@@ -8,7 +8,7 @@ import torch
 from torch import nn
 
 from ..base import EntityEmbeddingModel
-from ...hpo.search_spaces import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
+from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
 from ...losses import Loss
 from ...regularizers import Regularizer
 from ...triples import TriplesFactory
