+++ src/pykeen/models/unimodal/structured_embedding.py
@@ -12,7 +12,7 @@ from torch import nn
 from torch.nn import functional
 
 from ..base import EntityEmbeddingModel
-from ...hpo.search_spaces import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
+from ...constants import DEFAULT_EMBEDDING_HPO_EMBEDDING_DIM_RANGE
 from ...losses import Loss
 from ...nn import Embedding
 from ...nn.init import xavier_uniform_
